Badly needed patch for graphics package (GWTCanvas and ImageLoader). Much Improved IE implementation. Performance improvements (via fast string appends) and pixel for pixel correctness with native canvas (thanks to Oliver Zoran for his help).
Also layed groundwork for Gradients. They work in all browsers except IE, but the framework is there for a contributor to pick up and finish.

Fixed some bugs with ImageLoader and moved over to use the ImageElement class in the new DOM api instead of the old custom image handle.

Added new demos.

+++ src/com/google/gwt/demos/gwtcanvas/client/ParticleDemo.java
@@ -90,14 +90,14 @@ public class ParticleDemo extends SimpleCanvasDemo {
    * Not so much controls as feedback for benchmarking.
    */
   class ParticleDemoControls extends Composite {
-    private long average = 1;
+    private double average = 1;
     private Label averageLabel;
     
-    private long iterations = 1;
-    private long refreshRate;
+    private double iterations = 1;
+    private double refreshRate;
     private Label refreshRateLabel;
     
-    private long startTime = -1;
+    private double startTime = -1;
     
     public ParticleDemoControls() {
       refreshRateLabel = new Label("");
@@ -110,7 +110,7 @@ public class ParticleDemo extends SimpleCanvasDemo {
       initWidget(layout);
     }
     
-    public void doBenchmark(long now) {
+    public void doBenchmark(double now) {
       if (startTime < 0) {
         startTime = now;
       } else {
@@ -130,7 +130,9 @@ public class ParticleDemo extends SimpleCanvasDemo {
       this.startTime = -1;
     } 
   }
+  
   private int numParticles = 20;
+  
   private Particle[] particles;
   
   private Timer t;
